Fixed broken unit tests
Fixed broken unit tests. The tests were reading JSON input files that were using an out-of-date schema. Updating the schema fixed the tests. Specifically, the new schema uses a flat structure for ethnicity while the old one assumed ethnicity could hold more than one value. In the use case were this code would have been deployed, the data producers would only be able to select one ethnicity. The use of a JSON array for ethnicity caused the validator to see the field holding the ethnicity array and assumed there should have been a corresponding __code field, and when it did not find it, it threw an error. The additional errors failed assertions made in the unit tests.